backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.27",
    "django-cors-headers>=4.9.0",
    "django-crontab>=0.7.1",
    "djangorestframework>=3.16.1",
    "google-genai>=1.47.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.13",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.8",
]
